requirements.txt
asgiref==3.8.1
Django==5.2
django-appconf==1.1.0
django-imagekit==5.0.0
pilkit==3.0
pillow==11.1.0
python-dotenv==1.1.0
sqlparse==0.5.3
